djangodrf/pyproject.toml
[project]
name = "python-api-practices-examples-djangodrf"
version = "0.1.0"
description = "Django/DRF-track examples for Python API Practices — Vol. 1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Django>=4.2",
  "djangorestframework>=3.15",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-django>=4.8",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
DJANGO_SETTINGS_MODULE = "app.settings"
pythonpath = ["app"]
